Renaming module-name
