+++ command-client/build.gradle.kts
@@ -1,7 +1,6 @@
 val liquibaseVersion = rootProject.properties["liquibaseVersion"] as String
 val kotestVersion = rootProject.properties["kotestVersion"] as String
 val slf4jVersion = rootProject.properties["slf4jVersion"] as String
-val kotlinVersion = rootProject.properties["kotlinVersion"] as String
 
 dependencies {
     // liquibase
@@ -12,8 +11,6 @@ dependencies {
     // test
     testImplementation("io.kotest:kotest-framework-engine-jvm:$kotestVersion")
     testImplementation("io.kotest:kotest-runner-junit5:$kotestVersion")
-    // reflection
-    testImplementation("org.jetbrains.kotlin:kotlin-reflect:$kotlinVersion")
 }
 
 tasks.test {
